pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tz-weblog"
version = "0.0.1"
description = "static site generator from markdown files."
authors = [{ name="tzbits", email="tzbits@proton.me" }]
license = { file = "LICENSE" }
dependencies = [
  "PyYAML",
  "markdown",
  "click",
]
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/tzbits/tz-workspace"

[project.scripts]
weblog = "tz_weblog.main:cli"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]
